+++ core/errors.go
@@ -152,6 +152,15 @@ func (f *TemplateParseError) Error() string {
 	return fmt.Sprintf("failed to parse %s", f.Msg)
 }
 
+type ExecError struct {
+	Err      error
+	ExitCode int
+}
+
+func (e *ExecError) Error() string {
+	return ""
+}
+
 func CheckIfError(err error) {
 	if err != nil {
 		Exit(err)
@@ -159,11 +168,14 @@ func CheckIfError(err error) {
 }
 
 func Exit(err error) {
-	switch err.(type) {
+	switch err := err.(type) {
 	case *ConfigErr:
 		// Errors are already mapped with `error:` prefix
 		fmt.Fprintf(os.Stderr, "%v", err)
 		os.Exit(1)
+	case *ExecError:
+		// Don't print anything when there's a ExecError: server execution failed
+		os.Exit(err.ExitCode)
 	default:
 		fmt.Fprintf(os.Stderr, "%s: %v\n", text.FgRed.Sprintf("error"), err)
 		os.Exit(1)