+++ core/run/text.go
@@ -5,10 +5,12 @@ import (
 	"bytes"
 	"fmt"
 	"github.com/jedib0t/go-pretty/v6/text"
+	"golang.org/x/crypto/ssh"
 	"golang.org/x/exp/slices"
 	"golang.org/x/term"
 	"io"
 	"os"
+	"os/exec"
 	"strings"
 	"sync"
 	"text/template"
@@ -29,7 +31,7 @@ func (run *Run) Text(dryRun bool) error {
 		if run.Task.Spec.Parallel {
 			go func(i int, wg *sync.WaitGroup) {
 				defer wg.Done()
-				// TODO
+				// TODO: Handle errors when running tasks in parallel
 				_ = run.TextWork(i, prefixMaxLen, dryRun)
 			}(i, &wg)
 		} else {
@@ -39,16 +41,24 @@ func (run *Run) Text(dryRun bool) error {
 				return err
 			}(i, &wg)
 
-			switch err.(type) {
-			case *template.ExecError:
-				return err
-			case *core.TemplateParseError:
-				return err
-			default:
-				if run.Task.Spec.AnyErrorsFatal && err != nil {
-					// The error is printed for each server in method RunTextCmd.
-					// We just return early so other tasks are not executed.
-					return nil
+			if err != nil {
+				switch err.(type) {
+				case *template.ExecError:
+					return err
+				case *core.TemplateParseError:
+					return err
+				default:
+					if run.Task.Spec.AnyErrorsFatal {
+						// Return proper exit code for failed tasks
+						switch err := err.(type) {
+						case *ssh.ExitError:
+							return &core.ExecError{Err: err, ExitCode: err.ExitStatus()}
+						case *exec.ExitError:
+							return &core.ExecError{Err: err, ExitCode: err.ExitCode()}
+						default:
+							return err
+						}
+					}
 				}
 			}
 		}
@@ -102,7 +112,7 @@ func (run *Run) TextWork(rIndex int, prefixMaxLen int, dryRun bool) error {
 		case *core.TemplateParseError:
 			return err
 		default:
-			if err != nil && !task.Spec.IgnoreErrors {
+			if !task.Spec.IgnoreErrors && err != nil {
 				return err
 			}
 		}